refactor: utils 폴더 생성, constants.js 파일명 오타 수정
